+++ src/main/java/com/hanshin/supernova/auth/v2/dto/request/SecurityAuthLoginRequest.java
@@ -1,0 +1,10 @@
+package com.hanshin.supernova.auth.v2.dto.request;
+
+import lombok.Data;
+
+@Data
+public class SecurityAuthLoginRequest {
+
+    private String email;
+    private String password;
+}
