+++ src/main/java/com/hanshin/supernova/auth/v2/dto/response/SecurityAuthLoginResponse.java
@@ -1,0 +1,12 @@
+package com.hanshin.supernova.auth.v2.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class SecurityAuthLoginResponse {
+
+    private String nickname;
+    private String token;
+}
