+++ src/main/java/com/hanshin/supernova/security/config/WebMvcConfig.java
@@ -1,0 +1,21 @@
+package com.hanshin.supernova.security.config;
+
+import org.springframework.boot.web.servlet.view.MustacheViewResolver;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewResolverRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    @Override
+    public void configureViewResolvers(ViewResolverRegistry registry) {
+        MustacheViewResolver resolver = new MustacheViewResolver();
+        resolver.setCharset("UTF-8");
+        resolver.setContentType("text/html;charset=UTF-8");
+        resolver.setPrefix("classpath:/templates/");
+        resolver.setSuffix(".html");
+
+        registry.viewResolver(resolver);
+    }
+}
