--- src/main/java/com/hanshin/supernova/user/infrastructure/UserRepository.java
@@ -1,24 +1,0 @@
-package com.hanshin.supernova.user.infrastructure;
-
-import com.hanshin.supernova.user.domain.Authority;
-import com.hanshin.supernova.user.domain.User;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
-
-@Repository
-@Transactional(readOnly = true)
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    boolean existsByEmail(String email);
-
-    boolean existsByNickname(String nickname);
-
-    Optional<User> findByEmail(String email);
-
-    Optional<User> findByAuthority(Authority authority);
-}
