+++ src/main/java/com/hanshin/supernova/user/v2/domain/SecurityUser.java
@@ -1,0 +1,38 @@
+package com.hanshin.supernova.user.v2.domain;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Email;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Builder
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class SecurityUser {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Email
+    @Column(name = "email", unique = true)
+    private String email;
+
+    private String password;
+
+    private String username;
+
+    @Column(name = "nickname", unique = true)
+    private String nickname;
+
+//    private String role; // ROLE_USER, ROLE_ADMIN
+    @Enumerated
+    private Authority authority;
+
+    @Embedded
+    private Activity activity;
+}
